Refactor EthosnAPI compilation functions into EthosnCompiler
